move models and dtos to shared package
