+++ apps/api/go.mod
@@ -4,6 +4,11 @@ go 1.23.0
 
 toolchain go1.23.7
 
+
+require TimeTrack-shared v0.0.0
+
+replace TimeTrack-shared => ../../shared
+
 require (
 	github.com/gin-gonic/gin v1.10.1
 	github.com/swaggo/files v1.0.1
