+++ shared/go.mod
@@ -1,0 +1,3 @@
+module TimeTrack-shared
+
+go 1.23.0
